Create I2C_Bus_Sharing.c
It's the source code of I2C_Bus_Sharing of these embedded system. In these purpose, I2C_Done() function prevents that another sensor from who having the bus, uses the bus,  until it finishes its work in this system tour (transmits the necessary information to the microprocessor and receives the necessary answers). This prevents the sensors from interrupting each other's communication. If the bus is idle, the function returns true so that the next sensor receives the bus. So all sensors use the bus in sequence.